Initial commit of md2pdf-client.py
This is the first version of md2pdf-client that is available to the
world - there of course has been a fair amount of work before this point
in messing around and trying things out, but that's not really relevant
anymore.
Also, I realise that this is effectively useless without the
corresponding md2pdf-webserver application - I will make a separate git
repository for that some time soon, it still needs a little more QA
before I'm ready to release it.