Improve argparse configuration
The argparse config now requires either '--file' or '--set-default' to
be passed as an argument, and these are mutually exclusive. The idea is
to use the '--set-default' option to write to a simple config file in
the user directory so that this file does not need to be manually edited
when changes are needed.