Implement 'compare mode'
Refs #15
"Compare Mode" is intended to be used to highlight the differences
between two versions of a Markdown document. It is activated by passing
an extra parameter to `md2pdf-client`, the `--compare` flag, along with
a filename. `md2pdf-client` will then strip all images from both input
files (and replace them with a LaTeX fbox containing just the caption),
and create a ZIP for uploading to the server just like it does now.

This commit also refactors parts of the code, to improve readability and
reduce copy-pasting.