Support reading YAML metadata
Adds a simple handler to read the YAML metadata block at the beginning
of an input Markdown file.
Using the new parser, the template is now automatically passed through
the headers, meaning a custom template can now be designated in the
YAML metadata instead of using the command-line flag. Refs #5.